Modified cpl.graph to comply with the .clang-tidy rules (#81)
Modified cpl.graph to comply with the .clang-tidy rules.

Now, all cpl modules have been updated according to the .clang-tidy rules. No warning is generated for now, at least with clang-tidy 3.8.

Additionaly, the .travis.yml file has been modified to update the gcc version from 4.9 to 5 because there were some problems installing the 4.9 version in the machine. Anyway, it is a good idea to update the compiler.